Move parse path related to separate file in config package
